+++ lib/location/location_bloc.dart
@@ -4,6 +4,7 @@ import 'package:bloc/bloc.dart';
 import 'package:sendrax/location/location_event.dart';
 import 'package:sendrax/location/location_state.dart';
 import 'package:sendrax/models/climb.dart';
+import 'package:sendrax/models/grade_repo.dart';
 import 'package:sendrax/models/location.dart';
 import 'package:sendrax/models/location_repo.dart';
 import 'package:sendrax/models/user.dart';
@@ -17,8 +18,11 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
 
   final SelectedLocation location;
   final List<String> categories;
+  StreamController filterGradeStream = StreamController<String>.broadcast();
+  StreamController filterSectionStream = StreamController<String>.broadcast();
   StreamSubscription<List<Climb>> climbSubscription;
   StreamSubscription<Location> locationSubscription;
+  StreamSubscription<List<String>> gradesSubscription;
 
   @override
   LocationState get initialState {
@@ -31,6 +35,7 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
     final user = await UserRepo.getInstance().getCurrentUser();
     _retrieveClimbsForThisLocation(user);
     _retrieveSectionsForThisLocation(user);
+    _retrieveGradesForThisLocation(user);
   }
 
   void _retrieveClimbsForThisLocation(User user) async {
@@ -55,6 +60,27 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
     }
   }
 
+  void _retrieveGradesForThisLocation(User user) async {
+    if (user != null) {
+      gradesSubscription =
+          GradeRepo.getInstance().getGradesForId(user, location.gradesId).listen((grades) {
+        add(GradesUpdatedEvent(grades));
+      });
+    } else {
+      add(LocationErrorEvent());
+    }
+  }
+
+  void selectGrade(String grade) {
+    state.filterGrade = grade;
+    filterGradeStream.add(grade);
+  }
+
+  void selectSection(String section) {
+    state.filterSection = section;
+    filterSectionStream.add(section);
+  }
+
   @override
   Stream<LocationState> mapEventToState(LocationEvent event) async* {
     if (event is ClearDataEvent) {
@@ -63,6 +89,8 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
       yield LocationState.updateClimbs(false, event.climbs, state);
     } else if (event is SectionsUpdatedEvent) {
       yield LocationState.updateSections(false, event.sections, state);
+    } else if (event is GradesUpdatedEvent) {
+      yield LocationState.updateGrades(false, event.grades, state);
     } else if (event is LocationErrorEvent) {
       yield LocationState.loading(false, state);
     }
@@ -70,11 +98,16 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
 
   @override
   Future<void> close() {
+    filterGradeStream.close();
+    filterSectionStream.close();
     if (locationSubscription != null) {
       locationSubscription.cancel();
     }
     if (climbSubscription != null) {
       climbSubscription.cancel();
+    }
+    if (gradesSubscription != null) {
+      gradesSubscription.cancel();
     }
     return super.close();
   }