+++ lib/location/location_view.dart
@@ -75,31 +75,79 @@ class LocationWidget extends StatelessWidget {
                   textAlign: TextAlign.center,
                 ),
               );
-            } else if (state.sections.isNotEmpty) {
-              content = ListView.builder(
-                padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
-                itemBuilder: (context, index) {
-                  return _buildSection(context, state, index);
-                },
-                itemCount: state.sections.length,
-              );
             } else {
-              content = ListView.builder(
-                padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
-                itemBuilder: (context, index) {
-                  return _buildClimb(context, state, index);
-                },
-                itemCount: state.climbs.length,
-              );
+              content = _buildFilteredList(state, context);
             }
             return _wrapContentWithFab(state, context, content);
           }),
     );
   }
 
-  Widget _buildSection(BuildContext context, LocationState state, int index) {
-    List<Climb> climbsToInclude =
-        List.from(state.climbs.where((climb) => climb.section == state.sections[index]));
+  Widget _buildFilteredList(LocationState state, BuildContext context) {
+    return StreamBuilder(
+        stream: BlocProvider.of<LocationBloc>(context).filterGradeStream.stream,
+        initialData: state.filterGrade,
+        builder: (BuildContext context, gradeSnapshot) {
+          return StreamBuilder(
+              stream: BlocProvider.of<LocationBloc>(context).filterSectionStream.stream,
+              initialData: state.filterSection,
+              builder: (BuildContext context, sectionSnapshot) {
+                if (List.from(state.climbs.where((climb) =>
+                    (sectionSnapshot.data == null || climb.section == sectionSnapshot.data) &&
+                    (gradeSnapshot.data == null || climb.grade == gradeSnapshot.data))).isEmpty) {
+                  return _showEmptyFilteredList(state, context);
+                }
+                if (state.sections.isNotEmpty) {
+                  return _buildContentWithSections(
+                      state, context, sectionSnapshot.data, gradeSnapshot.data);
+                } else {
+                  return _buildContentWithoutSections(state, context, gradeSnapshot.data);
+                }
+              });
+        });
+  }
+
+  Widget _showEmptyFilteredList(LocationState state, BuildContext context) {
+    return ListView.builder(
+      padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
+      itemBuilder: (context, index) {
+        if (index == 0) {
+          return _showFilterDropdownRow(state, context);
+        } else {
+          return Center(
+              child: Text(
+            "There are no climbs matching these filter parameters.",
+            textAlign: TextAlign.center,
+          ));
+        }
+      },
+      itemCount: 2,
+    );
+  }
+
+  Widget _buildContentWithSections(
+      LocationState state, BuildContext context, String sectionSnapshot, String gradeSnapshot) {
+    return ListView.builder(
+      padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
+      itemBuilder: (context, index) {
+        if (index == 0) {
+          return _showFilterDropdownRow(state, context);
+        } else {
+          if (sectionSnapshot != null) {
+            return _buildSection(context, state, gradeSnapshot, sectionSnapshot);
+          } else {
+            return _buildSection(context, state, gradeSnapshot, state.sections[index - 1]);
+          }
+        }
+      },
+      itemCount: (state.filterSection != null) ? 2 : state.sections.length + 1,
+    );
+  }
+
+  Widget _buildSection(
+      BuildContext context, LocationState state, String filterGrade, String section) {
+    List<Climb> climbsToInclude = List.from(state.climbs.where((climb) =>
+        climb.section == section && (filterGrade == null || climb.grade == filterGrade)));
     if (climbsToInclude.isNotEmpty) {
       return Column(children: <Widget>[
         Row(children: <Widget>[
@@ -107,7 +155,7 @@ class LocationWidget extends StatelessWidget {
             child: Container(
               color: Colors.pinkAccent,
               padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
-              child: Text(state.sections[index],
+              child: Text(section,
                   overflow: TextOverflow.ellipsis,
                   style: TextStyle(fontSize: UIConstants.BIGGER_FONT_SIZE, color: Colors.white)),
             ),
@@ -127,17 +175,123 @@ class LocationWidget extends StatelessWidget {
     }
   }
 
-  Widget _buildClimb(BuildContext context, LocationState state, int index) {
-    if (state.climbs.isNotEmpty) {
+  Widget _buildContentWithoutSections(
+      LocationState state, BuildContext context, String gradeSnapshot) {
+    List<Climb> climbsToInclude =
+        List.from(state.climbs.where((climb) => climb.grade == gradeSnapshot));
+    return ListView.builder(
+      padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
+      itemBuilder: (context, index) {
+        if (index == 0) {
+          return _showFilterDropdownRow(state, context);
+        } else {
+          return _buildClimb(context, state, climbsToInclude, index - 1);
+        }
+      },
+      itemCount: climbsToInclude.length + 1,
+    );
+  }
+
+  Widget _buildClimb(BuildContext context, LocationState state, List<Climb> climbs, int index) {
+    if (climbs.isNotEmpty) {
       return Column(children: [
         new InkWell(
-            child: _buildItem(state.climbs[index]),
+            child: _buildItem(climbs[index]),
             onTap: () {
-              navigateToClimb(state.climbs[index], state);
+              navigateToClimb(climbs[index], state);
             })
       ]);
     } else {
       return Column();
+    }
+  }
+
+  ClimbItem _buildItem(Climb climb) {
+    return ClimbItem(climb: climb);
+  }
+
+  Widget _showFilterDropdownRow(LocationState state, BuildContext context) {
+    return Row(children: <Widget>[
+      Expanded(
+        child: _showGradeDropdown(state, context),
+      ),
+      Expanded(
+        child: _showSectionDropdown(state, context),
+      )
+    ]);
+  }
+
+  Widget _showGradeDropdown(LocationState state, BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(UIConstants.SMALLER_PADDING, 0.0,
+          UIConstants.SMALLER_PADDING, UIConstants.BIGGER_PADDING),
+      child: new StreamBuilder(
+          stream: BlocProvider.of<LocationBloc>(context).filterGradeStream.stream,
+          initialData: state.filterGrade,
+          builder: (BuildContext context, snapshot) {
+            return new Row(children: <Widget>[
+              Expanded(
+                  child: DropdownButton<String>(
+                items: _createDropdownItems(state.grades),
+                value: snapshot.data,
+                hint: Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: <Widget>[
+                  Icon(Icons.filter_list, color: Colors.grey),
+                  Text("Grade"),
+                ]),
+                isExpanded: true,
+                onChanged: (value) => BlocProvider.of<LocationBloc>(context).selectGrade(value),
+              )),
+              IconButton(
+                  icon: Icon(Icons.cancel),
+                  onPressed: () => BlocProvider.of<LocationBloc>(context).selectGrade(null))
+            ]);
+          }),
+    );
+  }
+
+  Widget _showSectionDropdown(LocationState state, BuildContext context) {
+    return Padding(
+        padding: const EdgeInsets.fromLTRB(UIConstants.SMALLER_PADDING, 0.0,
+            UIConstants.SMALLER_PADDING, UIConstants.BIGGER_PADDING),
+        child: new StreamBuilder(
+          stream: BlocProvider.of<LocationBloc>(context).filterSectionStream.stream,
+          initialData: state.filterSection,
+          builder: (BuildContext context, snapshot) {
+            return new Row(children: <Widget>[
+              Expanded(
+                child: DropdownButton<String>(
+                  disabledHint: Text("No sections"),
+                  iconDisabledColor: Colors.grey,
+                  items: _createDropdownItems(state.sections),
+                  value: snapshot.data,
+                  hint: Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: <Widget>[
+                    Icon(Icons.filter_list, color: Colors.grey),
+                    Text("Section"),
+                  ]),
+                  isExpanded: true,
+                  onChanged: (value) => BlocProvider.of<LocationBloc>(context).selectSection(value),
+                ),
+              ),
+              IconButton(
+                  icon: Icon(Icons.cancel,
+                      color: (state.sections.isEmpty) ? Colors.grey : Colors.black),
+                  onPressed: () => BlocProvider.of<LocationBloc>(context).selectSection(null))
+            ]);
+          },
+        ));
+  }
+
+  List<DropdownMenuItem> _createDropdownItems(List<String> items) {
+    if (items.isNotEmpty) {
+      return items.map((String value) {
+        return new DropdownMenuItem<String>(
+          value: value,
+          child: new Text(value),
+        );
+      }).toList();
+    } else {
+      // null disables the dropdown
+      return null;
     }
   }
 
@@ -156,10 +310,6 @@ class LocationWidget extends StatelessWidget {
         )
       ],
     );
-  }
-
-  ClimbItem _buildItem(Climb climb) {
-    return ClimbItem(climb: climb);
   }
 
   void _createClimb(LocationState state) {